Added help + arg parsing

+++ sort.py
@@ -1,18 +1,25 @@
+import argparse
 from collections import OrderedDict
 import hashlib
 import json
 import os
 from pick import pick
+from playsound import playsound
 import re
 from zipfile import ZipFile
 
-SONGS = "/home/mesch/programming/song-sorter/songs"
-PLAYLISTS = "/home/mesch/programming/song-sorter/playlists"
-DOWNLOADS = "/home/mesch/Downloads"
+parser = argparse.ArgumentParser(description="Utility script for managing BeatSaber Custom songs.")
+parser.add_argument("--downloads-dir", metavar="<downloads-folder>", type=str, default="/home/mesch/Downloads", help="Root directory of Downloads folder. Where songs are unzipped from.", dest="DOWNLOADS_DIR")
+parser.add_argument("--songs-dir", metavar="<custom-songs-folder>", type=str, default="/home/mesch/programming/song-sorter/songs", help="Root directory of Custom Songs Folder. Where songs will be searched for and added to.", dest="SONGS_DIR")
+parser.add_argument("--playlists-dir", metavar="<playlists-folder>", type=str, default="/home/mesch/programming/song-sorter/playlists", help="Root directory of Song Playlists Folder. Uses rithik-b/PlaylistManager", dest="PLAYLISTS_DIR")
+parser.add_argument("--no-unzip", action='store_true', help="If flag is set, do not attempt to unzip new songs from downloads folder into Custom Songs Folder.", dest="NO_UNZIP")
+parser.add_argument("--no-sort", action='store_true', help="If flag is set, do not prompt user to add unsorted songs to playlists.", dest="NO_SORT")
+parser.add_argument("--play-song", action='store_true', help="If flag is set, play songs suring playlist selection.", dest="PLAY_SONG")
+args = parser.parse_args()
 
 def getZippedSongs():
     zipped_songs = []
-    for root, dirs, files in os.walk(DOWNLOADS):
+    for root, dirs, files in os.walk(args.DOWNLOADS_DIR):
         for zip_file in [os.path.join(root, f) for f in files if f.endswith('.zip')]:
             with ZipFile(zip_file, 'r') as zf:
                 if re.search("info.dat", " ".join(zf.namelist()), flags=re.IGNORECASE):
@@ -22,7 +29,7 @@ def getZippedSongs():
 
 def getPlaylists():
     playlists = []
-    for root, dirs, files in os.walk(PLAYLISTS):
+    for root, dirs, files in os.walk(args.PLAYLISTS_DIR):
         playlists.extend([os.path.join(root, f) for f in files if f.endswith('.bplist')])
     return playlists
 
@@ -45,17 +52,19 @@ def generateSongHash(info_file, song_data):
     return hashlib.sha1(hash_string.encode()).hexdigest()
 
 def getSongInfo():
-    song_list = {}
-    for song_dir in [e for e in os.scandir(SONGS) if e.is_dir()]:
+    song_info = {}
+    for song_dir in [e for e in os.scandir(args.SONGS_DIR) if e.is_dir()]:
         reg = re.search("info.dat", " ".join(os.listdir(song_dir)), flags=re.IGNORECASE)
         if reg:
             song_info_path = os.path.join(song_dir.path, reg.group(0))
             with open(song_info_path, encoding="utf8") as f:
                 song_data = json.load(f, object_pairs_hook=OrderedDict)
                 song_hash = generateSongHash(song_info_path, song_data)
-                song_list[song_hash.upper()] = song_data["_songName"]
+                song_info[song_hash.upper()] = {"SongName": song_data["_songName"],
+                                                "ArtistName": song_data["_songAuthorName"],
+                                                "SongFile": os.path.join(song_dir.path, song_data["_songFilename"])}
     
-    return song_list
+    return song_info
 
 def addSongToPlaylist(song_name, song_hash, playlist_file):
     with open(playlist_file, 'r+', encoding="utf8") as f:
@@ -66,19 +75,28 @@ def addSongToPlaylist(song_name, song_hash, playlist_file):
         f.write(json.dumps(playlist_data, indent=4))
         f.truncate()
 
-zipped  = getZippedSongs()
-hashed = getHashedSongs()
-songs = getSongInfo()
+def unzipNewSongs():
+    zipped = getZippedSongs()
+    for zip_file in zipped:
+        unzip_loc = os.path.join(args.SONGS_DIR, os.path.splitext(os.path.basename(zip_file))[0])
+        if not os.path.exists(unzip_loc):        
+            with ZipFile(zip_file, 'r') as zf:
+                zf.extractall(unzip_loc)
+            os.remove(zip_file)
 
-for zip_file in zipped:
-    with ZipFile(zip_file, 'r') as zf:
-        zf.extractall(os.path.join(SONGS, os.path.splitext(os.path.basename(zip_file))[0]))
-    os.remove(zip_file)
+def promptAddUnsortedToPlaylist(play_audio):
+    hashed = getHashedSongs()
+    songs = getSongInfo()
+    for s_hash,s_info in songs.items():
+        if s_hash not in hashed:
+            selection_title = f"Song '{s_info['SongName']}' by '{s_info['ArtistName']}' not in any playlist.\nAdd '{s_info['SongName']}' to playlist:"
+            selection_options = (["SKIP"] + [os.path.splitext(os.path.relpath(p, args.PLAYLISTS_DIR))[0] for p in getPlaylists()])
+            selection, idx = pick(selection_options, selection_title, indicator="-->")
+            if selection != "SKIP":
+                addSongToPlaylist(s_info["SongName"], s_hash, getPlaylists()[idx-1])
 
-for s_hash,s_title in songs.items():
-    if s_hash not in hashed:
-        selection_title = f"Song '{s_title}' not in any playlist.\nAdd '{s_title}' to playlist:"
-        selection_options = (["SKIP"] + [os.path.splitext(os.path.relpath(p, PLAYLISTS))[0] for p in getPlaylists()])
-        selection, idx = pick(selection_options, selection_title, indicator="-->")
-        if selection != "SKIP":
-            addSongToPlaylist(s_title, s_hash, getPlaylists()[idx-1])
+if not args.NO_UNZIP:
+    unzipNewSongs()
+
+if not args.NO_SORT:
+    promptAddUnsortedToPlaylist(args.PLAY_SONG)
